Address travis build issue (#43)
This PR is an attempt to address the most recent failing travis builds.

$ npm test
> brewcalc-ui@0.1.0 test /home/travis/build/brewcomputer/brewcalc/app
> react-scripts test --env=jsdom
There might be a problem with the project dependency tree.
It is likely not a bug in Create React App, but something you need to fix locally.
The react-scripts package provided by Create React App requires a dependency:
  "babel-jest": "23.6.0"
Don't try to install it manually: your package manager does it automatically.
However, a different version of babel-jest was detected higher up in the tree:
  /home/travis/build/brewcomputer/brewcalc/node_modules/babel-jest (version: 24.8.0)
The build issue appears to be the result of npm using the parent directory's node_modules when attempting to resolve dependencies for the app/ travis steps. Removing the node_modules folder from the project root after the main library has been built should prevent the errors seen in travis.

PS: Apologies for the random PR. I'm planning on using this library in a personal recipe building tool. I may be interested in adding a few new features to the library in the future and figured I'd start by addressing the small build issue.